fix: run Opera on Windows 8
This is strange. For the longest time I've been able to ask BrowserStack to run
Opera on Windows 10. However, lately when requesting Opera on Windows 10, Karma
hangs. I went to the BrowserStack page that lists browsers and platforms and
Opera is not listed as supported on Windows 10. ?? I'm switching the default to
Windows 8 so that we can continue testing on Opera.